# object-container/src/swagger copy.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Object Container
  description: The micro-service intended to provide the mechanisum to store the objects to the minio server.
  contact:
    email: prasanna@makonlinesolutions.com
host: localhost:5005
basePath: /
securityDefinitions:
  jwt:
    type: apiKey
    name: Authorization
    in: header
schemes:
  - http
  - https
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
paths:
...
  /bucket/create:
    x-swagger-router-controller: bucket.controller
    post:
      summary: recently modeified objects list
      description: get all recently updated or modeified files list
      operationId: bucketCreate
      parameters:
        - in: body
          name: Bucket object
          description: Bucket name
          schema:
            type: object
            required:
              - bucketName
            properties:
              bucketName:
                type: string
      consumes:
        - application/json
      responses:
        "200":
          description: Success
          schema:
            properties:
              id:
                type: integer
                format: int32
        "400":
          description: Error
          schema:
            properties:
              message:
                type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string